requirements.txt
streamlit
langchain-community
langchain-groq
langchain-core
pymysql
sqlalchemy
sqlparse
pandas
python-dotenv